requirements.txt
# core dependency
pydantic==2.12.5 
scapy==2.7.0 
typer==0.24.1 
dnspython==2.8.0
holehe==1.61
shodan==1.31.0
python-whois==0.9.6
python-dotenv>=1.0.0

# additional dependencies
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.13.0
attrs==26.1.0
beautifulsoup4==4.14.3
bs4==0.0.2
certifi==2026.2.25
charset-normalizer==3.4.6
click==8.3.1
click-plugins==1.1.1.2
colorama==0.4.6
filelock==3.25.2
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
markdown-it-py==4.0.0
mdurl==0.1.2
outcome==1.3.0.post0
pydantic_core==2.41.5
Pygments==2.19.2
python-dateutil==2.9.0.post0
requests==2.33.0
requests-file==3.0.1
rich==14.3.3
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.8.3
termcolor==3.3.0
tldextract==5.3.1
tqdm==4.67.3
trio==0.33.0
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3
xlsxwriter==3.2.9